add sendMail to notification kafka
