+++ notification-service/src/main/java/com/example/notificationservice/email/EmailSenderService.java
@@ -1,0 +1,27 @@
+package com.example.notificationservice.email;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+@Service
+public class EmailSenderService {
+    @Autowired
+    private JavaMailSender mailSender;
+
+    public void sendEmail(String toEmail,
+                          String subject,
+                          String body){
+        SimpleMailMessage message= new SimpleMailMessage();
+        message.setFrom("banatward@gmail.com");
+        message.setTo(toEmail);
+        message.setText(body);
+        message.setSubject(subject);
+
+        mailSender.send(message);
+
+        System.out.println("Mail Sent successfully...");
+
+    }
+}
