+++ notification-service/src/main/resources/application.yml
@@ -26,3 +26,16 @@ spring:
       ddl-auto: update
     database: postgresql
     database-platform: org.hibernate.dialect.PostgreSQLDialect
+  mail:
+    host: smtp.gmail.com
+    port: 587
+    username: banatward2017@gmail.com
+    password: iiupxbbvfkyshiwv
+
+    properties:
+      mail:
+        smtp:
+          auth: true
+          starttls:
+            enable: true
+
