+++ service-post/src/main/java/com/nabilaitnacer/servicepost/PostService.java
@@ -44,7 +44,6 @@ public class PostService {
         PostEntity postEntity = modelMapper.map(postEntityDto, PostEntity.class);
         postEntity = postRepository.save(postEntity);
         // kafka by s
-        System.out.println("***************postEvent by service Post *****"+new PostPlacedEvent(postEntity.getId().toString(),postEntity.getUserId().toString()));
         kafkaTemplate.send("notificationTopic", new PostPlacedEvent(postEntity.getId().toString(),postEntity.getUserId().toString()));
 
         PostResponse postResponse = new PostResponse();
